Also call modprobe directly from the unit file
This helps in the case where configuration is changed and the user
calls 'systemctl daemon-reload && systemctl start swap.target'. This
would "hang" if the zram module wasn't already loaded. Let's keep
both mechanisms to load the module in place: this has higher chance
of working if e.g. selinux is active. If systemd-modules-load.service
does its job as expected, this additional modprobe call is a noop.

Hopefully fixes #4.