Add a real option parser and rework how config is handled
Let's just drive the execution of the program directly from main().
This allows the logic to be split:
- main: handling of arguments and environment, control of execution
- config: parsing of the config

This will make it much easier to add unit tests later on.
We also get --help and --version, which is occasionally useful.
I'll follow up with a commit that further simplifies config.