Drop authors list from Cargo.toml
clap shows the list in --help, which is not very useful.

And the list is going to be always out of date. So let's just list a few
names in README, and redirect people to the live list on github.

Also remove (c) line from the LICENSE file. It is not necessary for anything,
everything is copyrighted by default.