Use pkg-config for systemd directories
Add pkg-config calls to the Makefile.

Use an environment variable to pass the location of systemd-makefs to
the Rust code. This gets evaluated at compile-time. If the env var is
unset (e.g. when cargo is invoked by hand), a default is used.

The unit file also needs to be edited before installation.