units: rename swap-create@.service to systemd-zram-setup@.service
In subsequent patches I want to reuse the same service to set up
non-swap devices. Continuing to use the name "swap-create" would
be confusing.

"zram-setup@" would be OK, but udisk2 also had zram-setup@.service,
and renamed it to udisk2-zram-setup@.service. To avoid confusion,
let's add the "systemd-" prefix to show that this is from under the
systemd project umbrealla.