units: split out BindsTo= line to a generated dropin
In subsequent commits, systemd-zram-setup@.service will be used
for non-swap purposes, so the BindsTo= line needs to become optional.

Unit descriptions are changed to use titlecase. This is the general
recommended systemd style, because those strings are used in messages
like "Starting ..." and "Stopping ...", and a non-capitalized string
looks strange.