mount-options: Add Options= Setting to Set Mount Options
When using zram for file system, we need to enable `discard` mount
option for issue discard/TRIM commands to the underlying zram block
device when blocks are freed. Else deleted files are not zero-ed, so
zram does not remove the compressed pages.

For example, setting `/etc/systemd/zram-generator.conf` as below could
format the file system as ext4 with TRIM enabled:

    $ cat /etc/systemd/zram-generator.conf
    [zram11]
    mount-point = /var/compressed
    fs-type = ext4
    mount-options = discard

Fixes #89

Signed-off-by: Wong Hoi Sing Edison <hswong3i@pantarei-design.com>