make: allow install target that doesn't build anything
'make install NOBUILD=1' will just do the installation without trying to build
first.

This is useful when we want to split the build and installation steps, and if
there is anything to build in the installation step it is an error. For
example, this would be case when doing 'make build && sudo make install'.