Replace clap with getopts
Version 2 is really old. The version 3 update is easy,
but updating to version 4 is really complicated:
https://github.com/systemd/zram-generator/pull/208

We don't need anything that clap gives us, and it's a fight uphill both
ways to keep up with it and not pull in the universe.

This parser is trivial, and results in a fully-150k-smaller binary:

With clap:
   $ l target/release/zram-generator
   -rwxr-xr-x 2 nabijaczleweli users 762.5k 11-27 15:40 target/release/zram-generator
   $ size target/release/zram-generator
      text    data     bss     dec     hex filename
    740420   30177     392  770989   bc3ad target/release/zram-generator

Without:
   $ l target/release/zram-generator
   -rwxr-xr-x 2 nabijaczleweli users 598.5k 11-27 16:17 target/release/zram-generator
   $ size target/release/zram-generator
      text    data     bss     dec     hex filename
    577136   24457     392  601985   92f81 target/release/zram-generator

(After strip --strip-all.)