Cargo.toml: update clap version to 4
This changes the organization of the arguments and strictly speaking is an
incompatible change. Before, the device argument could be specified before
--reset-device or --setup-device, but now it must be after the option name.
I think this is an acceptable change, because this isn't really a public
interface and this is the natural way to specify this. This way the setup
is simpler and easier to understand at a glance.

I added #[rustfmt::skip] because the formatter insists on squishing some
arg descriptions into a single line which doesn't look good and also
indents the modifiers after '--setup-device' and 'dir' options differently
for some reason.

[#[rustfmt::skip] moved, attributes on expressions are experimental error]
[<String>+.clone() -> <&str>+.to_string() removes a copy]